Убрал дублирующий код и поместил его в отдельный класc ErrorHandler
