+++ src/main/java/ru/yandex/practicum/filmorate/controller/ErrorResponse.java
@@ -1,4 +1,19 @@
 package ru.yandex.practicum.filmorate.controller;
 
-public record ErrorResponse(String error, String description) {
+public class ErrorResponse {
+    private final String error;
+    private final String description;
+
+    public ErrorResponse(String error, String description) {
+        this.error = error;
+        this.description = description;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public String getDescription() {
+        return description;
+    }
 }
