+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -29,22 +29,4 @@ public class GenreController {
     public Genre getFilmById(@PathVariable int id) {
         return genreService.getById(id);
     }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse validationHandle(ValidationException e) {
-        return new ErrorResponse("error", "Указаны некорректные данные.");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse notFoundHandle(NullPointerException e) {
-        return new ErrorResponse("error", "Не найдено.");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_EXTENDED)
-    public ErrorResponse exceptionHandle(RuntimeException e) {
-        return new ErrorResponse("error", "Ошибка сервера.");
-    }
 }
