+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -67,29 +67,4 @@ public class UserController {
     public void removeFromFriendList(@PathVariable long id, @PathVariable(value = "friend-id") long friendId) {
         userService.deleteUserFromFriend(id, friendId);
     }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse validationHandle(ValidationException e) {
-        return new ErrorResponse("error", "Указаны некорректные данные.");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse notFoundHandle(NullPointerException e) {
-        return new ErrorResponse("error", "Не найдено.");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ErrorResponse notFoundHandle(NoSuchElementException e) {
-        return new ErrorResponse("error", "Элемент не найден.");
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_EXTENDED)
-    public ErrorResponse exceptionHandle(RuntimeException e) {
-        System.out.println(e.getLocalizedMessage());
-        return new ErrorResponse("error", "Ошибка сервера.");
-    }
 }
